linuxkm/lkcapi_sha_glue.c:
* add error percolation to wc_linuxkm_rng_registry_add_rng() and wc_linuxkm_rng_registry_add_bank(), and treat failures as fatal in wc_linuxkm_rng_bank_init() and linuxkm_InitRng_DefaultRBGC().

* bikeshed wc_mix_pool_bytes() with staggered-startup-offset.

wolfcrypt/src/rng_bank.c:

* add missing backward-compat WC_DRBG_NOT_INIT and WC_DRBG_OK definitions.

* add missing old-old-FIPS wc_RNG_DRBG_Reseed_Now() prototype and wc_RNG_DRBG_Stir() implementation.

wolfssl/wolfcrypt/rng_bank.h: fix sign-clash from WC_RNG_BANK_OFFSET_TO_RNG().

+++ wolfssl/wolfcrypt/rng_bank.h
@@ -386,7 +386,7 @@ WOLFSSL_API int wc_rng_new_bankref(struct wc_rng_bank *bank, WC_RNG **rng);
 #define WC_RNG_BANK_INST_TO_RNG(rng_inst) \
     ((rng_inst) ? (&(rng_inst)->rng) : NULL)
 #define WC_RNG_BANK_OFFSET_TO_RNG(bank, n) \
-    ((((n) >= 0) && ((unsigned)(n) < (unsigned)(bank)->n_rngs)) ? \
+    ((((int)(n) >= 0) && ((int)(n) < (int)(bank)->n_rngs)) ? \
      WC_RNG_BANK_INST_TO_RNG(&(bank)->rngs[n]) : NULL)
 
 #ifdef WC_RNG_HAVE_LOCK
